+++ src/foundation/exceptions/Handler/drivers/ViewDriver.js
@@ -23,7 +23,11 @@ class ViewDriver extends Driver {
 	 * @inheritdoc
 	 */
 	render(exception, request, response) {
-		response.end(this.renderView(exception));
+		if (this.wantsJson(request)) {
+			return response.json(this.renderJson(exception));
+		}
+
+		return response.end(this.renderView(exception));
 	}
 
 	/**