+++ test/unit/support/services/Dumper.test.js
@@ -10,6 +10,7 @@ beforeEach(() => {
 	given.providersAndDependencies();
 	given.bootedContainer();
 	given.emptyResult();
+	given.emptyFileName();
 	given.fakeTerminal();
 	given.fakeViewFactory();
 	given.baseDevConfig();
@@ -54,4 +55,46 @@ test('Can dump circular object', () => {
 	given.circularData();
 	when.dumping();
 	then.shouldNotHaveThrown();
+});
+
+test('Cannot dump if in a forbidden environment silently', () => {
+	given.disabledTestEnvironmentInConfiguration();
+	given.data();
+	given.fakeResponse();
+	when.dumping();
+	then.shouldNotHaveDumpedInResponse();
+	then.shouldNotHaveDumpedInConsole();
+	then.shouldNotHaveRendered();
+});
+
+test('Cannot get dump if in a forbidden environment silently', () => {
+	given.disabledTestEnvironmentInConfiguration();
+	given.data();
+	given.fakeResponse();
+	when.gettingDump();
+	then.shouldNotHaveDumpedInResponse();
+	then.shouldNotHaveDumpedInConsole();
+	then.shouldNotHaveRendered();
+});
+
+test('Can dump with a specific file name', () => {
+	given.data();
+	given.fakeResponse();
+	given.fileName();
+	when.dumpingForFile();
+	then.shouldHaveDumpedInResponse();
+	then.shouldHaveUsedGivenFileName();
+	then.shouldNotHaveDumpedInConsole();
+});
+
+test('Can get dump with a specific file name', () => {
+	given.data();
+	given.fakeResponse();
+	given.fileName();
+	when.gettingDumpForFile();
+	then.shouldNotHaveDumpedInResponse();
+	then.shouldNotHaveDumpedInConsole();
+	then.shouldHaveReceivedDumpRenderedView();
+	then.shouldHaveRenderedHtmlPartial();
+	then.shouldHaveUsedGivenFileName();
 });